pyproject.toml
[project]
name = "flask-demo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "flask>=3.1.1",
    "flask-sqlalchemy>=3.1.1",
    "pillow>=11.3.0",
    "werkzeug>=3.1.3",
]
